+++ flowman-common/src/main/scala/com/dimajix/flowman/common/Logging.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Kaya Kupferschmidt
+ * Copyright 2020-2022 Kaya Kupferschmidt
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,32 +26,50 @@ import java.util.Properties
 import scala.collection.JavaConverters._
 
 import org.apache.log4j.PropertyConfigurator
+import org.apache.logging.log4j.Level
+import org.apache.logging.log4j.LogManager
+import org.apache.logging.log4j.core.LoggerContext
+import org.apache.logging.log4j.core.config.DefaultConfiguration
+import org.apache.logging.log4j.core.{Logger => Log4jLogger}
 import org.slf4j.LoggerFactory
+import org.slf4j.impl.StaticLoggerBinder
 
 
 class Logging
 object Logging {
     private lazy val logger = LoggerFactory.getLogger(classOf[Logging])
-    private var props:Properties = null
+    private var log4j1Properties:Properties = null
     private var level:String = null
 
     def init() : Unit = {
-        val loader = Thread.currentThread.getContextClassLoader
-        val log4j = System.getProperty("log4j.configuration")
-        val url =
-            if (log4j == null || log4j.isEmpty) {
-                loader.getResource("com/dimajix/flowman/log4j-defaults.properties")
-            } else {
-                try {
-                    new URL(log4j)
+        val log4j1Config = System.getProperty("log4j.configuration")
+        val log4j2Config = System.getProperty("log4j.configurationFile")
+        if (isLog4j2() && log4j2Config != null) {
+            // Do nothing, config is already correctly loaded
+        }
+        else if (log4j1Config != null) {
+            val url = try {
+                    new URL(log4j1Config)
                 }
                 catch {
-                    case _:MalformedURLException => new File(log4j).toURI.toURL
+                    case _: MalformedURLException => new File(log4j1Config).toURI.toURL
                 }
-            }
-        props = loadProperties(url)
+            initlog4j1(url)
+        }
+        else {
+            initLog4jDefault()
+        }
+    }
+    private def initLog4jDefault() : Unit = {
+        val loader = Thread.currentThread.getContextClassLoader
+        val url = loader.getResource("com/dimajix/flowman/log4j-defaults.properties")
+        initlog4j1(url)
+
+    }
+    private def initlog4j1(configUrl:URL) : Unit = {
+        log4j1Properties = loadProperties(configUrl)
         reconfigureLogging()
-        logger.debug(s"Loaded logging configuration from $url")
+        logger.debug(s"Loaded logging configuration from $configUrl")
     }
 
     private def loadProperties(url:URL) : Properties = {
@@ -94,9 +112,9 @@ object Logging {
 
     private def reconfigureLogging() : Unit = {
         // Create new logging properties with all levels being replaced
-        val newProps = new Properties()
-        if (props != null) {
-            props.asScala.foreach { case (k, v) => newProps.setProperty(k, v) }
+        if (log4j1Properties != null) {
+            val newProps = new Properties()
+            log4j1Properties.asScala.foreach { case (k, v) => newProps.setProperty(k, v) }
             if (level != null) {
                 newProps.keys().asScala.collect { case s: String => s }
                     .toList
@@ -105,10 +123,37 @@ object Logging {
             }
             PropertyConfigurator.configure(newProps)
         }
+        else if (isLog4j2()) {
+            if (level != null) {
+                val context = LogManager.getContext(false).asInstanceOf[LoggerContext]
+                val l = Level.getLevel(level)
+                val config = context.getConfiguration
+                config.getRootLogger.setLevel(l)
+                config.getLoggers.asScala.values.foreach(_.setLevel(l))
+                context.updateLoggers()
+            }
+        }
 
         if (level != null) {
             val l = org.apache.log4j.Level.toLevel(level)
             org.apache.log4j.Logger.getRootLogger().setLevel(l)
         }
+    }
+
+    private def isLog4j2(): Boolean = {
+        // This distinguishes the log4j 1.2 binding, currently
+        // org.slf4j.impl.Log4jLoggerFactory, from the log4j 2.0 binding, currently
+        // org.apache.logging.slf4j.Log4jLoggerFactory
+        val binderClass = StaticLoggerBinder.getSingleton.getLoggerFactoryClassStr
+        "org.apache.logging.slf4j.Log4jLoggerFactory".equals(binderClass)
+    }
+
+    private def isLog4j2DefaultConfigured(): Boolean = {
+        val rootLogger = LogManager.getRootLogger.asInstanceOf[Log4jLogger]
+        rootLogger.getAppenders.isEmpty ||
+            (rootLogger.getAppenders.size() == 1 &&
+                rootLogger.getLevel == Level.ERROR &&
+                LogManager.getContext.asInstanceOf[LoggerContext]
+                    .getConfiguration.isInstanceOf[DefaultConfiguration])
     }
 }